Use pg_control_checkpoint() instead of pg_control file parsing
