+++ cmd/wal-g/main.go
@@ -300,7 +300,7 @@ func main() {
 		if err != nil {
 			log.Fatalf("%+v\n", err)
 		}
-		n, err := bundle.StartBackup(conn, time.Now().String(), dirArc)
+		n, err := bundle.StartBackup(conn, time.Now().String())
 		if err != nil {
 			log.Fatalf("%+v\n", err)
 		}
@@ -336,7 +336,7 @@ func main() {
 			log.Fatalf("%+v\n", err)
 		}
 
-		timelineChanged := bundle.CheckTimelineChanged(dirArc)
+		timelineChanged := bundle.CheckTimelineChanged(conn)
 
 		// Wait for all uploads to finish.
 		err = bundle.Tb.Finish(!timelineChanged)
